#EJC20231026: Corregí el tipo de dato de posicion X y posición Y

+++ app/src/main/java/com/dts/classes/Mantenimientos/Bodega/clsBeBodega_ubicacion.java
@@ -45,9 +45,9 @@ public class clsBeBodega_ubicacion {
     @Element(required=false) public String NombreCompleto="";
     @Element(required=false) public double Disponibilidad_Ubicacion=0;
 
-    @Element(required=false) public int Posicion_X = 0;
+    @Element(required=false) public double Posicion_X = 0;
 
-    @Element(required=false) public int Posicion_Y = 0;
+    @Element(required=false) public double Posicion_Y = 0;
 
     public clsBeBodega_ubicacion() {
     }
@@ -61,7 +61,7 @@ public class clsBeBodega_ubicacion {
                                  boolean Ubicacion_picking,boolean Ubicacion_recepcion,boolean Ubicacion_despacho,boolean Ubicacion_merma,
                                  boolean Ubicacion_Virtual,double Margen_izquierdo,double Margen_derecho,double Margen_superior,
                                  double Margen_inferior,String Orientacion_pos,boolean ubicacion_ne,clsBeBodega_tramo Tramo,
-                                 clsBeBodega_sector Sector,String NombreCompleto, int Posicion_X, int Posicion_Y) {
+                                 clsBeBodega_sector Sector,String NombreCompleto, double Posicion_X, double Posicion_Y) {
 
         this.IdUbicacion=IdUbicacion;
         this.IdTramo=IdTramo;
@@ -343,17 +343,17 @@ public class clsBeBodega_ubicacion {
         Disponibilidad_Ubicacion=value;
     }
 
-    public int getPosicion_X() {
+    public double getPosicion_X() {
         return Posicion_X;
     }
-    public void setPosicion_X(int value) {
+    public void setPosicion_X(double value) {
         Posicion_X=value;
     }
 
-    public int getPosicion_Y() {
+    public double getPosicion_Y() {
         return Posicion_Y;
     }
-    public void setPosicion_Y(int value) {
+    public void setPosicion_Y(double value) {
         Posicion_Y=value;
     }
